Update: Replace route blueprints with functions
