+++ packages/app/app/router.js
@@ -4,6 +4,9 @@
  */
 import EmberRouter from '@ember/routing/router';
 import config from './config/environment';
+import { reportRoutes, reportCollectionRoutes, reportPrintRoutes } from 'navi-reports/router';
+import { dashboardRoutes, dashboardCollectionRoutes, dashboardPrintRoutes } from 'navi-dashboards/router';
+import { directoryRoutes } from 'navi-directory/router';
 
 const Router = EmberRouter.extend({
   location: config.locationType,
@@ -12,44 +15,13 @@ const Router = EmberRouter.extend({
 
 Router.map(function() {
   this.route('landing', { path: '/' });
-
-  this.route('directory', function() {
-    this.route('my-data');
-  });
-
-  this.route('dashboard-collections', function() {
-    this.route('collection', { path: '/:collection_id' });
-  });
-
-  this.route('dashboards', function() {
-    this.route('new');
-    this.route('dashboard', { path: '/:dashboard_id' }, function() {
-      this.route('clone');
-      this.route('view');
-
-      this.route('widgets', function() {
-        this.route('add');
-        this.route('new');
-        this.route('widget', { path: '/:widget_id' }, function() {
-          this.route('edit');
-          this.route('invalid');
-          this.route('view');
-          this.route('clone-to-report');
-        });
-      });
-    });
-  });
-
-  this.route('reports', function() {
-    this.route('new');
-    this.route('report', { path: '/:report_id' }, function() {
-      this.route('edit');
-      this.route('invalid');
-      this.route('view');
-      this.route('clone');
-      this.route('save-as');
-    });
-  });
+  directoryRoutes(this);
+  reportRoutes(this);
+  reportCollectionRoutes(this);
+  reportPrintRoutes(this);
+  dashboardRoutes(this);
+  dashboardCollectionRoutes(this);
+  dashboardPrintRoutes(this);
 });
 
 export default Router;
