+++ packages/dashboards/app/routes/dashboards-print/dashboards/dashboard/index.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-dashboards/routes/dashboards-print/dashboards/dashboard/index';
