--- packages/dashboards/app/routes/print/dashboards/dashboard/index.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-dashboards/routes/print/dashboards/dashboard/index';
