+++ packages/dashboards/app/templates/dashboards-print/dashboards/dashboard/view.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-dashboards/templates/dashboards-print/dashboards/dashboard/view';
