--- packages/dashboards/app/templates/print/dashboards/dashboard/view.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-dashboards/templates/print/dashboards/dashboard/view';
