+++ packages/dashboards/tests/dummy/app/router.js
@@ -1,5 +1,7 @@
 import EmberRouter from '@ember/routing/router';
 import config from './config/environment';
+import { reportRoutes } from 'navi-reports/router';
+import { dashboardRoutes, dashboardCollectionRoutes, dashboardPrintRoutes } from 'navi-dashboards/router';
 
 const Router = EmberRouter.extend({
   location: config.locationType,
@@ -7,48 +9,14 @@ const Router = EmberRouter.extend({
 });
 
 Router.map(function() {
+  //mock directoy route
   this.route('directory', function() {
     this.route('my-data');
   });
-
-  this.route('dashboard-collections', function() {
-    this.route('collection', { path: '/:collection_id' });
-  });
-  this.route('dashboards', function() {
-    this.route('new');
-
-    this.route('dashboard', { path: '/:dashboard_id' }, function() {
-      this.route('clone');
-      this.route('view');
-
-      this.route('widgets', function() {
-        this.route('add');
-        this.route('new');
-        this.route('widget', { path: '/:widget_id' }, function() {
-          this.route('clone-to-report');
-          this.route('edit');
-          this.route('view');
-          this.route('invalid');
-        });
-      });
-    });
-  });
-
-  this.route('print', function() {
-    this.route('dashboards', function() {
-      this.route('dashboard', { path: '/:dashboard_id' }, function() {
-        this.route('view');
-      });
-    });
-  });
-
-  this.route('reports', function() {
-    this.route('report', { path: '/:report_id' }, function() {
-      this.route('view');
-      this.route('clone');
-      this.route('edit');
-    });
-  });
+  reportRoutes(this);
+  dashboardRoutes(this);
+  dashboardCollectionRoutes(this);
+  dashboardPrintRoutes(this);
 });
 
 export default Router;
