--- packages/reports/app/controllers/print/reports/report.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-reports/controllers/print/reports/report';
