+++ packages/reports/app/controllers/reports-print/reports/report.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-reports/controllers/reports-print/reports/report';
