--- packages/reports/app/routes/print/reports/new.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-reports/routes/print/reports/new';
