--- packages/reports/app/routes/print/reports/report/index.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-reports/routes/print/reports/report/index';
