+++ packages/reports/app/routes/reports-print/reports/new.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-reports/routes/reports-print/reports/new';
