+++ packages/reports/app/routes/reports-print/reports/report/view.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-reports/routes/reports-print/reports/report/view';
