--- packages/reports/app/templates/print/reports/report.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-reports/templates/print/reports/report';
