--- packages/reports/app/templates/print/reports/report/view.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-reports/templates/print/reports/report/view';
