+++ packages/reports/app/templates/reports-print/reports/report.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-reports/templates/reports-print/reports/report';
